Is imported 'Hook' class in typescript example snippet
The example snippet is listed as a TypeScript file and it imports the Hook class, but it never uses is.

From looking at the example at the top of the docs it follows that it's a generic class that's intended to receive the hook name.  Based on the file name `post_analytics` I assume the hook would also be named `post-analytics' although since it's supposed to be a type and not just a string i'm not sure if this is correct.