Change link to "Build your own plugin" from Features to Plugins page
The "Build your own plugin" has a broken link to the features page which no longer contains this section. 
Fixing this to point to the Plugins page instead, and removing "below" from the text as it is no longer true.
The link is formatted according to other links in this document.